Added Live cam facility for users to directly input images from live cam

+++ core/home/models.py
@@ -5,7 +5,7 @@ import datetime
 class Students(models.Model):
     name = models.CharField(max_length=100)
 
-    roll_no = models.IntegerField(unique=True)
+    roll_no = models.BigIntegerField(unique=True)
 
     face_embeddings = models.BinaryField()
